* cleaned up installation code: * made range and trackingRange of installationTemplate int (float is confusing because radar_t use int) * don't save range and trackingRange: they are parsed in the ufo file. * don't check radar range for aircraft that are not on geoscape. * BREAKS SAVEGAME. If you want to keep your savegames, follow instructions below* first part of the commit: saving part. If you want to keep your current savegame, update your checkout to this revision, load your game and save it. Then update your checkout to next revision. * Second part will follow soon

+++ src/client/cl_installation.h
@@ -50,8 +50,8 @@ typedef struct installationTemplate_s {
 	char *name;		/**< Name of the installation (as you see it ingame). */
 
 	int cost;				/**< Price of the installation. */
-	float radarRange;		/**< The range of the installation's radar.  Units is the angle of the two points from center of earth. */
-	float trackingRange;	/**< The tracking range of the installation's radar. Units are degrees. */
+	int radarRange;			/**< The range of the installation's radar.  Units is the angle of the two points from center of earth. */
+	int trackingRange;		/**< The tracking range of the installation's radar. Units are degrees. */
 	int maxBatteries;		/**< The maximum number of battery slots that can be used in an installation. */
 	int maxUfoStored;		/**< The maximum number of ufos that can be stored in an installation. */
 	int maxDamage;			/**< The maximum amount of damage an installation can sustain before it is destroyed. */
